Merged PR 339: Fixed issue that caused the web driver not to be properly disposed after each scenario
Fixed issue that caused the web driver not to be properly disposed after each scenario.

A better fix would be to change the Browser property getter to create a new instance if `browser == null || browser.IsDisposed`. The `IsDisposed` property is protected - we could implement our own `Browser` class to get around this.