fix: GetUser returning wrong users for alias (#89)
* fix: GetUser(string alias) not returning new users

The default behaviour wasn't returning a user for the given alias when `currentUser` was `true`. It would also cycle users when requesting the same alias in a single scenario with `currentUser` was set to `true` if the calls were separated by a call for another alias.

* fix: exceptions creating base profiles

Depending on the test runner, this code would fail. Synchronisation also needs to happen across processes in the case where build agents are running multiple test runs concurrently.

* fix: CurrentUsers occasionally `null`

Co-authored-by: Max Ewing <max.ewing@outlook.com>